feat: adicionei as responsesEntities para os controllers

+++ src/main/java/dev/java10x/projetoapi/Ninjas/NinjaController.java
@@ -1,5 +1,8 @@
 package dev.java10x.projetoapi.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,33 +17,47 @@ public class NinjaController {
         this.ninjaService = ninjaService;
     }
 
-    @GetMapping("/boasvindas")
-    public String boasVindas(){
-        return "Boas-vindas - Olá mundo!";
-    }
-
     @PostMapping("/criar")
-    public NinjaDTO criarNinja(@RequestBody NinjaDTO ninja){
-        return ninjaService.criarNinja(ninja);
+    public ResponseEntity<String> criarNinja(@RequestBody NinjaDTO ninja) {
+        NinjaDTO ninjaDTO = ninjaService.criarNinja(ninja);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Ninja criado com sucesso: " + ninjaDTO.getNome());
     }
 
-    @GetMapping("/all")
-    public List<NinjaDTO> listarNinjas(){
-        return ninjaService.listarNinjas();
+    @GetMapping("/todos")
+    public ResponseEntity<List<NinjaDTO>>  listarNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.listarNinjas();
+        return ResponseEntity.status(HttpStatus.OK).body(ninjas);
     }
 
     @GetMapping("/todos/{id}")
-    public NinjaDTO mostrarNinjasID(@PathVariable Long id){
-        return ninjaService.listarNinjaPorId(id);
+    public ResponseEntity<?> mostrarNinjasID(@PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.listarNinjaPorId(id);
+
+        if(ninja != null){
+            return ResponseEntity.status(HttpStatus.OK).body(ninja);
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não foi possível encontrar o ninja com id fornecido");
     }
 
     @PutMapping("/alterar/{id}")
-    public NinjaDTO alterarNinjasID(@PathVariable Long id, @RequestBody NinjaDTO ninjaModel){
-        return ninjaService.atualizarNinja(id, ninjaModel);
+    public ResponseEntity<String> alterarNinjasID(@PathVariable Long id, @RequestBody NinjaDTO ninjaModel) {
+        NinjaDTO ninja = ninjaService.atualizarNinja(id, ninjaModel);
+
+        if(ninja != null){
+            return ResponseEntity.ok("Ninja alterado com sucesso");
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja não encontrado");
     }
 
     @DeleteMapping("/deletar/{id}")
-    public void deletarNinjas(@PathVariable Long id){
-        ninjaService.removerNinja(id);
+    public ResponseEntity<String> deletarNinjas(@PathVariable Long id) {
+        if (ninjaService.listarNinjaPorId(id) != null) {
+            ninjaService.removerNinja(id);
+            return ResponseEntity.ok("Ninja deletado com sucesso");
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja não encontrado");
     }
 }
